chore: move affinity functions to utils
